+++ Source/map.cpp
@@ -4,10 +4,6 @@
 #include "config.h"
 #include "tick.h"
 #include "globals.h"
-/* TODO (GamerMan7799#9#): Have Monster and player movements based off of velocity instead of set cells. Every
-                           update would calculate how far they go based on time that has passed (like I did for the Cannon Simulation).
-                           Inputs will add / subtract velocities. I will first have to figure out how to detect where tile bounding boxes are.
-                           It will be a lot of work will likely be part of the next major release v5.0.0*/
 /**********************************************************************************************************************************************/
 clsMap::clsMap() {
 	if (Global::blnDebugMode) {printf("Map Constructor called.\n");}
@@ -35,6 +31,8 @@ void clsMap::restart(void) {
 		for (uchar i = 0; i < numMonsters; i++) {
 			tempMonster.location.x = pmstBaseMonsters[i].location.x;
 			tempMonster.location.y = pmstBaseMonsters[i].location.y;
+			tempMonster.vel.x = pmstBaseMonsters[i].vel.x;
+			tempMonster.vel.y = pmstBaseMonsters[i].vel.y;
 			tempMonster.living = pmstBaseMonsters[i].living;
 			tempMonster.state = pmstBaseMonsters[i].state;
 			Global::Enty.setMonster(i,tempMonster);
@@ -42,47 +40,94 @@ void clsMap::restart(void) {
 		}
 		if (Global::blnDebugMode) {printf("Base monsters placed.\n");}
 	} else {if (Global::blnDebugMode) {printf("Base monsters equals nullptr.\n");}}
-	Global::Enty.setPlayer(locBasePlayer);
+	Global::Enty.setPlayer(bplyBasePlayer);
 	if (Global::blnDebugMode) {printf("Player reset.\n");}
 }
 /**********************************************************************************************************************************************/
 char clsMap::move(uchar direction) {
 	static uchar jumpcount; //how many times the player has jumped
 	static bool playerfalling; //if the player is falling
-	uchar tempx;
-	uchar tempy;
+	float deltat = 1.0 / DEFINED_GOAL_FPS;
+	BPLYR tempEntity; //keeps track of x, y, and velocity.
 	if (playerfalling == false) {jumpcount = 0;}
 
+	float GravityVelDec = Global::Physics::fGravityAcceleration * deltat;
+    bool blnLoop = true;
 	//Move monsters first, to see if player dies and we can then skip the rest
-
 	MNSTR tempMonster;
 
 	for (uchar i = 0; i < numMonsters; i++) {
 		tempMonster = Global::Enty.getMonster(i);
 		if (tempMonster.living) {
-			tempx = tempMonster.location.x;
-			tempy = tempMonster.location.y;
-			if (tempMonster.state == stateRight) {tempx ++;}
-			else if (tempMonster.state == stateLeft) {tempx--;}
+			tempEntity.location.x = tempMonster.location.x;
+			tempEntity.location.y = tempMonster.location.y;
+			tempEntity.vel.x = tempMonster.vel.x;
+			tempEntity.vel.y = tempMonster.vel.y;
 
-			if (map[tempy][tempx] == tileSpace) {
-				map[tempy][tempMonster.location.x] = tileSpace;
-				map[tempy][tempx] = tileMonster;
-				tempMonster.location.x = tempx;
-			}
-			else if (map[tempy][tempx] == tilePlayer) {return deathMonster;}
-			else if (map[tempy][tempx] == tileWall) {
-                if (tempMonster.state == stateRight) {tempMonster.state = stateLeft;}
-                else if (tempMonster.state == stateLeft) {tempMonster.state = stateRight;}
+            //Decrease velocities by friction
+            tempEntity.vel.x -= tempEntity.vel.x * Global::Physics::fFriction;
+            tempEntity.vel.y -= tempEntity.vel.y * Global::Physics::fFriction;
+
+			if (tempMonster.state == stateRight) {
+                if (tempEntity.vel.x + Global::Physics::fIncVelocity < Global::Physics::fMaxVelocity) {tempEntity.vel.x += Global::Physics::fIncVelocity;}
+                else {tempEntity.vel.x = Global::Physics::fMaxVelocity;}
+            } else if (tempMonster.state == stateLeft) {
+                if ( abs(tempEntity.vel.x - Global::Physics::fIncVelocity) < Global::Physics::fMaxVelocity) {tempEntity.vel.x -= Global::Physics::fIncVelocity;}
+                else {tempEntity.vel.x = -1.00 * Global::Physics::fMaxVelocity;}
             }
 
-			//Because I don't want to have to deal with monsters falling, if the space below a monster is space
-			//kill the monster.
-			if (map[tempy+1][tempMonster.location.x] == tileSpace ) {tempMonster.living = false;}
-			if(tempMonster.living == false) {map[tempMonster.location.y][tempMonster.location.x] = tileSpace;}
-		} //end if living
-		Global::Enty.setMonster(i,tempMonster);
-	} //End of for monsters
+            //Apply Gravity
+            if (tempEntity.vel.y + GravityVelDec < Global::Physics::fMaxVelocity) {tempEntity.vel.y += GravityVelDec;}
+            else {tempEntity.vel.y = Global::Physics::fMaxVelocity;}
+
+            //Update location based on delta t
+            tempEntity.location.x += round(tempEntity.vel.x * deltat);
+            tempEntity.location.y += round(tempEntity.vel.y * deltat);
+
+            //Check for Collisions
+            do {
+                switch ( checkCollision(tempEntity.location, tileMonster) ) {
+                case tileCoin:
+                    //Replace the coin tile with space
+                    map[(uint)(tempMonster.location.y / DEFINED_PIC_SIZE)][(uint)(tempMonster.location.x / DEFINED_PIC_SIZE)] = tileSpace;
+                case tilePole:
+                case tileSpace:
+                    tempMonster.location.x = tempEntity.location.x;
+                    tempMonster.location.y = tempEntity.location.y;
+                    tempMonster.vel.x = tempEntity.vel.x;
+                    tempMonster.vel.y = tempEntity.vel.y;
+                    blnLoop = false;
+                    break;
+                case tilePlayer:
+                    //Monster runs into player, kill the player
+                    blnLoop = false;
+                    return deathMonster;
+                    break;
+                case tileWall:
+                default:
+                    //Collision with wall first reset y
+                    if (tempEntity.location.y != tempMonster.location.y) {
+                        //Change y and try again.
+                        tempEntity.location.y = tempMonster.location.y;
+                        tempEntity.vel.y = 0;
+                        blnLoop = true;
+                    } else if (tempEntity.location.x != tempMonster.location.x) {
+                        //Change x and try again.
+                        //recoil off the wall as well
+                        tempEntity.location.x = tempMonster.location.x;
+                        tempEntity.vel.x *= Global::Physics::fRecoil;
+                        blnLoop = true;
+                    } else { tempMonster.living = false; blnLoop = false; }
+                } //end switch check Collision
+            } while (blnLoop);
+            //do a check if monster is in the kill planes
+            if (tempMonster.location.y < 0 || tempMonster.location.x < 0 || tempMonster.location.y > DEFINED_MAP_HEIGHT * DEFINED_PIC_SIZE || tempMonster.location.x > DEFINED_MAP_WIDTH * DEFINED_PIC_SIZE) {
+                tempMonster.living = false;
+            }
+
+            Global::Enty.setMonster(i,tempMonster);
+		} // end if living
+    }//End of for monsters
 
 
 	//Now the player can move.
@@ -92,59 +137,104 @@ char clsMap::move(uchar direction) {
     //Kill Player if their fitness gets too low in hard mode
     if (tempPlayer.fitness < -2.5 && (Global::Cnfg.getvalues(cnfgHardMode) == 1) ) {return deathDecay;}
 
-	tempx = tempPlayer.location.x;
-	tempy = tempPlayer.location.y;
+    tempEntity.location = tempPlayer.location;
+    tempEntity.vel = tempPlayer.vel;
+
+    //Decrease velocities by friction
+    tempEntity.vel.x -= tempEntity.vel.x * Global::Physics::fFriction;
+    tempEntity.vel.y -= tempEntity.vel.y * Global::Physics::fFriction;
 
 	switch (direction) {
 		case dirLeft :
 		    tempPlayer.state = stateLeft;
-			if (tempx != 0) {tempx --;}
+		    if ( abs(tempEntity.vel.x - Global::Physics::fIncVelocity) < Global::Physics::fMaxVelocity) {tempEntity.vel.x -= Global::Physics::fIncVelocity;}
+		    else {tempEntity.vel.x = Global::Physics::fMaxVelocity;}
 			break;
 		case dirRight :
 		    tempPlayer.state = stateRight;
-			if (tempx != DEFINED_MAP_WIDTH - 1) {tempx ++;}
+            if ( tempEntity.vel.x + Global::Physics::fIncVelocity < Global::Physics::fMaxVelocity) {tempEntity.vel.x += Global::Physics::fIncVelocity;}
+		    else {tempEntity.vel.x = Global::Physics::fMaxVelocity;}
 			break;
 		case dirUp :
-			if (map[tempy][tempx] == tileWall) {break;}
-			tempPlayer.state = stateJump;
-			//make sure that the player has enough space to make the jump or the program will attempt to reference a
-			//non-existing array spot (crashing the program). Also note that jumping is consider in negative direction because the top of the
-			//map is considered array spot 0 while the bottom is 13.
-			if (tempy > DEFINED_JUMP_HEIGHT && jumpcount < DEFINED_MAX_JUMP_COUNT) {tempy -= DEFINED_JUMP_HEIGHT; jumpcount++;}
+		    if (jumpcount < DEFINED_MAX_JUMP_COUNT) {
+                tempPlayer.state = stateJump;
+                jumpcount++;
+                if ( abs(tempEntity.vel.y - Global::Physics::fIncVelocity) < Global::Physics::fMaxVelocity) {tempEntity.vel.y -= Global::Physics::fIncVelocity;}
+                else {tempEntity.vel.y = Global::Physics::fMaxVelocity;}
+		    } //end if less than jump count
 			break;
+        case dirDown :
+            tempPlayer.state = stateDuck;
+            break;
+        case dirNone:
+        default:
+            tempPlayer.state = stateRest;
+            break;
 	};
-	if (tempPlayer.location.y == DEFINED_MAP_HEIGHT - 1) {return deathFall;}//This makes the very last row of the array a "kill plane"
 
-	if (playerfalling) {tempy++;}
-		if (map[tempy][tempx] == tileMonster) {
-		//If the player falls on a monster kill the monster,
-		//but if the player is trying to walk into a monster kill the player
-		if (playerfalling) {Global::Enty.killMonster(tempx,tempy); tempPlayer.score += DEFINED_MONS_KILL_POINTS;}
-		else {return deathMonster;}
-	}
+    //Apply Gravity
+    if (tempEntity.vel.y + GravityVelDec < Global::Physics::fMaxVelocity) {tempEntity.vel.y += GravityVelDec;}
+    else {tempEntity.vel.y = Global::Physics::fMaxVelocity;}
 
-	if (map[tempy][tempx] == tileWall){
-		//This checks a few spots for an open area they can move into if they run into a wall
-		if (map[tempy][tempPlayer.location.x] != tileWall) {tempx = tempPlayer.location.x;}
-		else if (map[tempPlayer.location.y][tempx] != tileWall) {tempy = tempPlayer.location.y;}
-		else if (map[tempy + 1][tempx] != tileWall) {tempy += 1;}
-		else {tempy = tempPlayer.location.y; tempx = tempPlayer.location.x;}
-	}
+    //Update location based on delta t
+    tempEntity.location.x += round(tempEntity.vel.x * deltat);
+    tempEntity.location.y += round(tempEntity.vel.y * deltat);
 
-	if (map[tempy][tempx] == tileCoin) {tempPlayer.score += DEFINED_COIN_WORTH;}
-	if (tempx < locBasePlayer.x - 2 && (Global::Cnfg.getvalues(cnfgHardMode) == 1)) {return deathStupid;} //if the player goes too far to the left kill them.
+	blnLoop = true;
+	//Check collisions
+	do {
+        switch ( checkCollision(tempEntity.location, tilePlayer) ) {
+        case tileCoin:
+            //Replace the coin tile with space
+            map[(uint)(tempMonster.location.y / DEFINED_PIC_SIZE)][(uint)(tempMonster.location.x / DEFINED_PIC_SIZE)] = tileSpace;
+            tempPlayer.score += DEFINED_COIN_WORTH;
+        case tilePole:
+        case tileSpace:
+            tempPlayer.location.x = tempEntity.location.x;
+            tempPlayer.location.y = tempEntity.location.y;
+            tempPlayer.vel.x = tempEntity.vel.x;
+            tempPlayer.vel.y = tempEntity.vel.y;
+            if (tempEntity.vel.y > 0) {playerfalling = true;}
+            blnLoop = false;
+            break;
+        case tileMonster:
+            //Player is running into monster
+            if (tempEntity.vel.y > 0) { //kill monster if falling on it
+                Global::Enty.killMonster(tempEntity.location.x, tempEntity.location.y);
+                tempPlayer.score += DEFINED_MONS_KILL_POINTS;
+            } else { return deathMonster;}
+            blnLoop = true;
+            break;
+        case tileWall:
+        default:
+            //Collision with wall first reset y
+            if (tempEntity.location.y != tempPlayer.location.y) {
+                //Change y and try again.
+                tempEntity.location.y = tempPlayer.location.y;
+                tempEntity.vel.y = 0.0;
+                playerfalling = false;
+                blnLoop = true;
+            } else if (tempEntity.location.x != tempPlayer.location.x) {
+                //Change x and try again.
+                //recoil off the wall as well
+                tempEntity.location.x = tempPlayer.location.x;
+                tempEntity.vel.x *= Global::Physics::fRecoil;
+                blnLoop = true;
+            } else { blnLoop = false; return deathError; }
+        } //end switch check Collision
+	} while (blnLoop);
 
-	map[tempPlayer.location.y][tempPlayer.location.x] = tileSpace;
-	map[tempy][tempx] = tilePlayer;
-	tempPlayer.location.x = tempx;
-	tempPlayer.location.y = tempy;
+    //do a check if player is in the kill planes
+    if (tempPlayer.location.y > DEFINED_MAP_HEIGHT * DEFINED_PIC_SIZE) {
+        return deathFall;
+    }
 
-	Global::Enty.setPlayer(tempPlayer.location);
-
-	if ((tempPlayer.location.y == DEFINED_MAP_HEIGHT - 1) || (map[tempPlayer.location.y + 1][tempPlayer.location.x] == tileSpace)) {
-        playerfalling = true;
-        tempPlayer.state = stateFalling;
-    } else {playerfalling = false;}
+    BPLYR temp;
+    temp.location = tempPlayer.location;
+    temp.vel = tempPlayer.vel;
+    //Check for stupid death (going too far left on hard mode)
+    if (tempPlayer.location.x < bplyBasePlayer.location.x - 2 * DEFINED_PIC_SIZE && (Global::Cnfg.getvalues(cnfgHardMode) == 1)) {return deathStupid;}
+    Global::Enty.setPlayer(temp);
 
 	//Reduce the clock time and check if it equals 0 and kill the player if it does.
 	Global::Tick.decClock();
@@ -157,16 +247,40 @@ void clsMap::load(void) {
 	//Finds player and monster on the map, and place them in base stats used
 	//when restarting the map.
 
+uchar tempmap[DEFINED_MAP_HEIGHT][DEFINED_MAP_WIDTH] = {{1,1,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,5,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,0,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,5,5,5,0,0,0,0},
+{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,5,0,0,0,0,0,0,0,5,5,5,5,5,1,0,0,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,1,1,1,0,0,0,0},
+{0,0,0,0,0,0,5,5,5,5,5,5,5,0,0,0,1,0,0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,1,0,0,0,0,0,1,0,0,1,0,0,1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,1,1,1,0,0,0,0},
+{0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,0,0,0,0,1,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,0,0,0,0,1,0,0,0,0,3,0,0,0,0,1,1,1,1,1,0,0,0},
+{0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,1,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,1,1,1,0,0,1,1,1,0,1,1,1,1,0,0,0,1,1,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,1,1,1,1,1,1,1,1,0,0,0,1,1,1,0,0,0,3,0,0,0,0,1,1,1,1,1,0,0,0},
+{1,1,1,1,1,2,0,0,0,0,0,0,0,1,1,1,1,1,1,0,0,0,4,0,0,1,1,1,1,0,0,0,0,0,0,0,4,0,0,0,0,1,1,0,1,1,1,1,0,1,1,1,1,1,1,1,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,1,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,1,1,1,1,1,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,0,0,3,0,0,0,1,1,1,1,1,1,0,0,0},
+{1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,1,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1},
+{1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1}};
+
+    //set base map (I couldn't do it as a const anymore because I'm trying to modify it)
+
     /* TODO (xPUREx#5#): Look into vectors for basemonsters */
 	numMonsters = 0;
 	for (uint y = 0; y < DEFINED_MAP_HEIGHT; y++) {
 		for (uint x = 0; x < DEFINED_MAP_WIDTH; x++) {
+            basemap[y][x] = tempmap[y][x];
 			if (basemap[y][x] == tilePlayer) {
 				if (Global::blnDebugMode) {printf("Found Player at (%d,%d).\n",x,y);}
-				locBasePlayer.x = x;
-				locBasePlayer.y = y;
+				bplyBasePlayer.location.x = x * DEFINED_PIC_SIZE;
+				bplyBasePlayer.location.y = y * DEFINED_PIC_SIZE;
+				bplyBasePlayer.vel.x = 0;
+				bplyBasePlayer.vel.y = 0;
+				basemap[y][x] = tileSpace; //Once we find the player they are going
+                                           //to be moving in a velocity based pattern
+                                           //so delete that tile
 			}
 			else if (basemap[y][x] == tileMonster) {
+                basemap[y][x] = tileSpace;
 				numMonsters++;
 				if (Global::blnDebugMode) {printf("Found Monster at (%d,%d)\n",x,y);}
 				if (numMonsters == 1) {
@@ -176,8 +290,10 @@ void clsMap::load(void) {
 						Global::blnError = true;
 						return;
 					}
-					pmstBaseMonsters[0].location.x = x;
-					pmstBaseMonsters[0].location.y = y;
+					pmstBaseMonsters[0].location.x = x * DEFINED_PIC_SIZE;
+					pmstBaseMonsters[0].location.y = y * DEFINED_PIC_SIZE;
+                    pmstBaseMonsters[0].vel.x = x;
+                    pmstBaseMonsters[0].vel.y = y;
 					pmstBaseMonsters[0].living = true;
 					pmstBaseMonsters[0].state = stateLeft;
 				} else {
@@ -195,6 +311,8 @@ void clsMap::load(void) {
 						//std::copy(pmstBaseMonsters, pmstBaseMonsters + numMonsters - 1, pTemp);
 						pTemp[i].location.x = pmstBaseMonsters[i].location.x;
 						pTemp[i].location.y = pmstBaseMonsters[i].location.y;
+						pTemp[i].vel.x = pmstBaseMonsters[i].vel.x;
+                        pTemp[i].vel.y = pmstBaseMonsters[i].vel.y;
 						pTemp[i].living = pmstBaseMonsters[i].living;
 						pTemp[i].state = pmstBaseMonsters[i].state;
 					}
@@ -202,8 +320,10 @@ void clsMap::load(void) {
 					delete[] pmstBaseMonsters;
 
 					//Place newest Monster values into temp array.
-					pTemp[numMonsters - 1].location.x = x;
-					pTemp[numMonsters - 1].location.y = y;
+					pTemp[numMonsters - 1].location.x = x * DEFINED_PIC_SIZE;
+					pTemp[numMonsters - 1].location.y = y * DEFINED_PIC_SIZE;
+                    pTemp[numMonsters - 1].vel.x = 0;
+                    pTemp[numMonsters - 1].vel.y = 0;
 					pTemp[numMonsters - 1].living = true;
 					pTemp[numMonsters - 1].state = stateLeft;
 
@@ -219,6 +339,8 @@ void clsMap::load(void) {
 						//Copy Temp Array into new array.
 						pmstBaseMonsters[i].location.x = pTemp[i].location.x;
 						pmstBaseMonsters[i].location.y = pTemp[i].location.y;
+						pmstBaseMonsters[i].vel.x = pTemp[i].vel.x;
+                        pmstBaseMonsters[i].vel.y = pTemp[i].vel.y;
 						pmstBaseMonsters[i].living = pTemp[i].living;
 						pmstBaseMonsters[i].state = pTemp[i].state;
 					}
@@ -244,7 +366,78 @@ void clsMap::setMapCell(uint x, uint y, uchar tile) {
 	map[y][x] = tile;
 }
 /**********************************************************************************************************************************************/
-LOC clsMap::getbasePlayer(void) {
-	return locBasePlayer;
+BPLYR clsMap::getbasePlayer(void) {
+	return bplyBasePlayer;
+}
+/**********************************************************************************************************************************************/
+char clsMap::checkCollision(LOC placement, uchar tile) {
+    //First convert pixel
+
+    BOX A;
+    BOX B;
+
+    A.left = placement.x;
+    A.right = placement.x + DEFINED_PIC_SIZE;
+    A.top = placement.y;
+    A.bottom = placement.y + DEFINED_PIC_SIZE;
+
+    //Check if collision with a player (if monster)
+    if (tile == tileMonster) {
+        PLYR tempPlayer = Global::Enty.getPlayer();
+        B.left = tempPlayer.location.x;
+        B.right = B.left + DEFINED_PIC_SIZE;
+        B.top = tempPlayer.location.y;
+        B.bottom = B.top + DEFINED_PIC_SIZE;
+
+        if ( checkOverlap(A, B) ) {
+            if (Global::blnDebugMode) {printf("Monster collided with player.\n");}
+            return tilePlayer;
+        } //end if overlap
+    } //end if monster
+
+    //Check if collision with a monster (if player)
+    if (tile == tilePlayer) {
+        MNSTR tempMonster;
+        for (uchar i = 0; i < numMonsters; i++) {
+            tempMonster = Global::Enty.getMonster(i);
+            B.left = tempMonster.location.x;
+            B.right = B.left + DEFINED_PIC_SIZE;
+            B.top = tempMonster.location.y;
+            B.bottom = B.top + DEFINED_PIC_SIZE;
+
+            if ( checkOverlap(A, B) ) {
+                if (Global::blnDebugMode) {printf("Player collided with monster.\n");}
+                return tileMonster;
+            } //end if overlap
+        } //end for monster
+    } //end if player
+
+
+
+    for (uint y = 0; y < DEFINED_MAP_HEIGHT; y++ ) {
+        for (uint x = 0; x < DEFINED_MAP_WIDTH; x++ ) {
+            if (map[y][x] != tileSpace) {
+                B.left = x * DEFINED_PIC_SIZE;
+                B.right = B.left + DEFINED_PIC_SIZE;
+                B.top = y * DEFINED_PIC_SIZE;
+                B.bottom = B.top + DEFINED_PIC_SIZE;
+                if ( checkOverlap(A, B) ) {
+                    if (Global::blnDebugMode) {printf("Map collision found.\n");}
+                    return map[y][x];
+                }
+            } //end if not space
+        } //end for x
+    } //end for y
+    return tileSpace; //return a space, for no collision
+}
+/**********************************************************************************************************************************************/
+bool clsMap::checkOverlap(BOX A, BOX B) {
+    if ( (A.right >= B.left && A.right <= B.right) || (A.left <= B.right && A.left >= B.left) ) { //X's cross now check y's
+        if ( (A.bottom < B.bottom && A.bottom > B.top) || (A.top > B.bottom && A.top < B.top) ) {
+            //if (Global::blnDebugMode) {printf("Overlap found.\n");}
+            return true;
+        } //end collide y
+    } //end collide x
+    return false;
 }
 /**********************************************************************************************************************************************/